modules/s8-python-network/s8-python-network/requirements.txt
aiohttp==3.7.4.post0
aio-pika==7.1.2
grpcio>=1.46.3
grpcio-health-checking~=1.46.3
grpcio-reflection~=1.46.3
grpcio-tools~=1.46.3
protobuf==3.19.4
psutil==5.9.0
pytest~=6.1.2
pytest-asyncio~=0.18.3
python-consul2==0.1.5
